ViewResolverBuilder package refactoring
