+++ src/main/java/com/agileandroid/ui/adapters/view/AgileAdapter.java
@@ -4,7 +4,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 
-import com.agileandroid.ui.adapters.view.builders.ViewResolverBuilder;
+import com.agileandroid.ui.adapters.view.resolver.ViewResolverBuilder;
 import com.agileandroid.ui.adapters.view.dto.AgileAdapterDTO;
 import com.agileandroid.ui.adapters.view.resolver.ViewResolver;
